Move WpfGraphControl into the project RealTimeGraphX.WPF
In my opinion, the control WpfGraphControl  belongs in the core DLL RealTimeGraphX.WPF and not in the demo project.
