+++ RealTimeGraphX.WPF.Demo/RealTimeGraphX.WPF.Demo.csproj
@@ -58,7 +58,6 @@
     </ApplicationDefinition>
     <Compile Include="Converters\ObjectEqualityConverter.cs" />
     <Compile Include="ListGraphItemVM.cs" />
-    <Compile Include="WpfGraphControl.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -71,10 +70,6 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Page Include="Themes\Generic.xaml">
-      <Generator>MSBuild:Compile</Generator>
-      <SubType>Designer</SubType>
-    </Page>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="MainWindowVM.cs" />
